code/requirements.txt
fastapi
uvicorn
httpx
pillow
pywebio